Update rustup and asdf scripts.

+++ rustup-bootstrap.sh
@@ -1,10 +1,15 @@
-#!/bin/sh
+#!/bin/bash
 set -e
 
-curl https://sh.rustup.rs -sSf | sh
-rustup install nightly
-rustup component add rust-src
-cargo install racer
-cargo install rustfmt
-cargo install ripgrep
-rustup run nightly cargo install clippy
+# Rust
+read -p "Do you want to install Rust? (y/n) " -n 1;
+echo "";
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+	curl https://sh.rustup.rs -sSf | sh -s -- --no-modify-path
+	rustup install nightly
+	rustup component add rust-src
+	cargo install racer
+	cargo install rustfmt
+	cargo install ripgrep
+	rustup run nightly cargo install clippy
+fi;
